add trending edit form to admin, add stok

+++ app/Http/Controllers/KeranjangController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Keranjang;
 use App\Models\Pesanan;
-use Illuminate\Support\Facades\DB;
+use App\Models\Produks;
+use App\Models\Keranjang;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class KeranjangController extends Controller
 {
@@ -28,19 +29,29 @@ class KeranjangController extends Controller
             foreach ($selectedItems as $itemId) {
                 $keranjangItem = Keranjang::find($itemId);
                 $pemesan = auth('siswa')->user()->id;
-                if ($keranjangItem) {
-                    Pesanan::create([
-                        'nama' => $keranjangItem->nama,
-                        'harga' => $keranjangItem->harga,
-                        'foto' => $keranjangItem->foto,
-                        'level' => $keranjangItem->level,
-                        'jumlah' => $keranjangItem->jumlah,
-                        'status' => 'Sedang diproses',
-                        'kantin_id' => $keranjangItem->kantin_id,
-                        'pemesan_id' => $pemesan
-                    ]);
 
-                    $keranjangItem->delete();
+                if ($keranjangItem) {
+
+                    $produk = Produks::where('nama', $keranjangItem->nama)->first();
+
+                    if ($produk && $produk->stok >= $keranjangItem->jumlah) {
+                        $produk->stok -= $keranjangItem->jumlah;
+                        $produk->save();
+                        Pesanan::create([
+                            'nama' => $keranjangItem->nama,
+                            'harga' => $keranjangItem->harga,
+                            'foto' => $keranjangItem->foto,
+                            'level' => $keranjangItem->level,
+                            'jumlah' => $keranjangItem->jumlah,
+                            'status' => 'Sedang diproses',
+                            'kantin_id' => $keranjangItem->kantin_id,
+                            'pemesan_id' => $pemesan
+                        ]);
+                        $keranjangItem->delete();
+                    } else {
+                        return redirect()->back()->with('error', 'Stok tidak cukup.');
+
+                    }
                 }
             }
         });
